test3/requirements.txt
python-Levenshtein
pymysql
requests
beautifulsoup4
lxml
